requirements.txt
fastapi
pdfminer.six
sentence_transformers 
pydantic 
numpy 
pinecone-client
python-dotenv
unicorn
